Move packaged asset provider into plugin tree
