+++ CMakeLists.txt
@@ -159,6 +159,7 @@ if(NOT EMSCRIPTEN)
 endif()
 add_subdirectory(plugin/obj_decoder)
 add_subdirectory(plugin/stl_decoder)
+add_subdirectory(plugin/resource_provider)
 
 add_subdirectory(src/engine)
 add_subdirectory(src/user)
