+++ src/user/user_resource.cc
@@ -17,25 +17,12 @@
 #include <sys/types.h>
 #include <sys/stat.h>
 
-#include <cctype>
 #include <cstddef>
 #include <cstdio>
-#include <cstdint>
-#include <cstdlib>
 #include <cstring>
 #include <ctime>
-#include <fstream>
-#include <mutex>
 #include <string>
 #include <string_view>
-#include <vector>
-
-#if defined(_WIN32) || defined(__CYGWIN__)
-#define NOMINMAX
-#include <windows.h>
-#elif !defined(__EMSCRIPTEN__)
-#include <dlfcn.h>
-#endif
 
 #include <mujoco/mujoco.h>
 
@@ -44,194 +31,8 @@
 #include "user/user_util.h"
 #include "user/user_vfs.h"
 
-namespace {
-
-constexpr char kMujocoAssetPrefix[] = "mujoco";
-constexpr char kMujocoAssetScheme[] = "mujoco:";
-constexpr char kMujocoAssetsDirEnv[] = "MUJOCO_ASSETS_DIR";
-constexpr char kMujocoFilamentAssetsDirEnv[] = "MUJOCO_FILAMENT_ASSETS_DIR";
-
-struct MujocoAssetResource {
-  std::vector<char> data;
-};
-
-bool IsMujocoAssetResourceName(const char* name) {
-  if (!name) {
-    return false;
-  }
-
-  for (int i = 0; kMujocoAssetScheme[i]; ++i) {
-    if (std::tolower(static_cast<unsigned char>(name[i])) !=
-        kMujocoAssetScheme[i]) {
-      return false;
-    }
-  }
-  return true;
-}
-
-std::string JoinPath(std::string_view dir, std::string_view filename) {
-  if (dir.empty()) {
-    return std::string(filename);
-  }
-  std::string path(dir);
-  const char last = path.back();
-  if (last != '/' && last != '\\') {
-#if defined(_WIN32) || defined(__CYGWIN__)
-    path.push_back('\\');
-#else
-    path.push_back('/');
-#endif
-  }
-  path.append(filename);
-  return path;
-}
-
-std::string GetLibraryDir() {
-#if defined(__EMSCRIPTEN__)
-  return "";
-#else
-  void* anchor = reinterpret_cast<void*>(
-      reinterpret_cast<uintptr_t>(&mj_forward));
-#if defined(_WIN32) || defined(__CYGWIN__)
-  HMODULE module = nullptr;
-  constexpr DWORD flags = GET_MODULE_HANDLE_EX_FLAG_FROM_ADDRESS |
-                          GET_MODULE_HANDLE_EX_FLAG_UNCHANGED_REFCOUNT;
-  if (GetModuleHandleExA(flags, reinterpret_cast<LPCSTR>(anchor), &module)) {
-    char path[MAX_PATH];
-    if (GetModuleFileNameA(module, path, MAX_PATH)) {
-      std::string lib_path(path);
-      const std::size_t last_slash = lib_path.find_last_of("\\/");
-      if (last_slash != std::string::npos) {
-        return lib_path.substr(0, last_slash + 1);
-      }
-    }
-  }
-#else
-  Dl_info info;
-  if (dladdr(anchor, &info) && info.dli_fname) {
-    std::string lib_path(info.dli_fname);
-    const std::size_t last_slash = lib_path.find_last_of('/');
-    if (last_slash != std::string::npos) {
-      return lib_path.substr(0, last_slash + 1);
-    }
-  }
-#endif
-  return "";
-#endif
-}
-
-std::string_view ResourceFilename(std::string_view name) {
-  const std::size_t prefix_end = name.find(':');
-  return prefix_end == std::string_view::npos ? name
-                                              : name.substr(prefix_end + 1);
-}
-
-bool IsSafeAssetFilename(std::string_view filename) {
-  if (filename.empty() || filename.front() == '/' || filename.front() == '\\') {
-    return false;
-  }
-
-  std::size_t segment_start = 0;
-  while (segment_start <= filename.size()) {
-    std::size_t segment_end = filename.find_first_of("/\\", segment_start);
-    if (segment_end == std::string_view::npos) {
-      segment_end = filename.size();
-    }
-    std::string_view segment =
-        filename.substr(segment_start, segment_end - segment_start);
-    if (segment == "..") {
-      return false;
-    }
-    segment_start = segment_end + 1;
-  }
-  return true;
-}
-
-int OpenMujocoAssetResource(mjResource* resource) {
-  const std::string_view filename = ResourceFilename(resource->name);
-  if (!IsSafeAssetFilename(filename)) {
-    return 0;
-  }
-
-  std::vector<std::string> asset_dirs;
-  if (const char* env_dir = std::getenv(kMujocoAssetsDirEnv)) {
-    asset_dirs.emplace_back(env_dir);
-  }
-  if (const char* env_dir = std::getenv(kMujocoFilamentAssetsDirEnv)) {
-    asset_dirs.emplace_back(env_dir);
-  }
-  const std::string library_dir = GetLibraryDir();
-  if (!library_dir.empty()) {
-    asset_dirs.push_back(JoinPath(library_dir, "assets"));
-  }
-  asset_dirs.emplace_back("assets");
-
-  for (const std::string& dir : asset_dirs) {
-    std::ifstream file(JoinPath(dir, filename),
-                       std::ios::binary | std::ios::ate);
-    if (!file.is_open()) {
-      continue;
-    }
-    const std::streamsize size = file.tellg();
-    if (size <= 0) {
-      continue;
-    }
-    auto* asset = new MujocoAssetResource();
-    asset->data.resize(static_cast<std::size_t>(size));
-    file.seekg(0, std::ios::beg);
-    if (!file.read(asset->data.data(), size)) {
-      delete asset;
-      continue;
-    }
-    resource->data = asset;
-    return static_cast<int>(asset->data.size());
-  }
-  return 0;
-}
-
-int ReadMujocoAssetResource(mjResource* resource, const void** buffer) {
-  auto* asset = static_cast<MujocoAssetResource*>(resource->data);
-  if (!asset) {
-    return -1;
-  }
-  *buffer = asset->data.data();
-  return static_cast<int>(asset->data.size());
-}
-
-void CloseMujocoAssetResource(mjResource* resource) {
-  delete static_cast<MujocoAssetResource*>(resource->data);
-  resource->data = nullptr;
-}
-
-void EnsureDefaultMujocoAssetProvider() {
-  static std::once_flag once;
-  std::call_once(once, [] {
-    if (mjp_getResourceProvider("mujoco:asset")) {
-      return;
-    }
-
-    mjpResourceProvider provider;
-    mjp_defaultResourceProvider(&provider);
-    provider.prefix = kMujocoAssetPrefix;
-    provider.open = OpenMujocoAssetResource;
-    provider.read = ReadMujocoAssetResource;
-    provider.close = CloseMujocoAssetResource;
-
-    const int slot = mjp_registerResourceProvider(&provider);
-    if (slot < 0 && !mjp_getResourceProvider("mujoco:asset")) {
-      mju_warning("Failed to register default MuJoCo asset resource provider.");
-    }
-  });
-}
-
-}  // namespace
-
 mjResource* mju_openResource(const char* dir, const char* name,
                              const mjVFS* vfs, char* error, size_t nerror) {
-  if (IsMujocoAssetResourceName(name)) {
-    EnsureDefaultMujocoAssetProvider();
-  }
-
   // TODO: Update API to use non-const pointer. Unfortunately, while this is
   // ABI stable, it will cause compiler errors in user code that is const
   // correct.
